+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - [Repository] Fetch remote data.
 - [Branches] Checkout remote branches locally.
-- [UX/UI] Resizeable left sidebar with persistant values. Double click to reset to the default value.
+- [UX/UI] Resizeable left sidebar with persistent values. Double click to reset to the default value.
 
 ## Version (1.0.12) (02/01/2026)