+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - [Branches] Checkout remote branches locally.
 - [UX/UI] Resizeable left sidebar with persistent values. Double click to reset to the default value.
 
+### Fixes
+
+- [Working directory / Commit history] Unselect file when committing, switching feature in sidebar.
+
 ## Version (1.0.12) (02/01/2026)
 
 ### Added
@@ -36,7 +40,7 @@
 
 ## Version (1.0.8) (30/12/2025)
 
-### Fix
+### Fixes
 
 - [App] Fix auto updater.
 
@@ -65,7 +69,7 @@
 
 - [Working directory] Discard all changes.
 
-### Fix
+### Fixes
 
 - [Branches] Fix branches remote error while creating local branch.
 
@@ -83,7 +87,7 @@
 
 ## Version (1.0.1) (28/12/2025)
 
-### Fix
+### Fixes
 
 - [Branches] Fix git branch checkout.